+++ pyccel/codegen/pipeline.py
@@ -159,7 +159,7 @@ def execute_pyccel(fname, *,
         folder = os.path.abspath(folder)
 
     # Define directory name and path for pyccel & cpython build
-    pyccel_dirname = '__pyccel__'
+    pyccel_dirname = '__pyccel__' + os.environ.get('PYTEST_XDIST_WORKER', '')
     pyccel_dirpath = os.path.join(folder, pyccel_dirname)
 
     # Create new directories if not existing
@@ -203,7 +203,8 @@ def execute_pyccel(fname, *,
     except NotImplementedError as error:
         msg = str(error)
         errors.report(msg+'\n'+PYCCEL_RESTRICTION_TODO,
-            severity='error')
+            severity='error',
+            traceback=error.__traceback__)
     except PyccelError:
         handle_error('parsing (syntax)')
         raise
@@ -222,7 +223,8 @@ def execute_pyccel(fname, *,
     except NotImplementedError as error:
         msg = str(error)
         errors.report(msg+'\n'+PYCCEL_RESTRICTION_TODO,
-            severity='error')
+            severity='error',
+            traceback=error.__traceback__)
     except PyccelError:
         handle_error('annotation (semantic)')
         # Raise a new error to avoid a large traceback
@@ -247,7 +249,8 @@ def execute_pyccel(fname, *,
     except NotImplementedError as error:
         msg = str(error)
         errors.report(msg+'\n'+PYCCEL_RESTRICTION_TODO,
-            severity='error')
+            severity='error',
+            traceback=error.__traceback__)
     except PyccelError:
         handle_error('code generation')
         # Raise a new error to avoid a large traceback
@@ -320,7 +323,7 @@ def execute_pyccel(fname, *,
     # ...
     # Determine all .o files and all folders needed by executable
     def get_module_dependencies(parser, deps):
-        mod_folder = os.path.join(os.path.dirname(parser.filename), "__pyccel__")
+        mod_folder = os.path.join(os.path.dirname(parser.filename), '__pyccel__' + os.environ.get('PYTEST_XDIST_WORKER', ''))
         mod_base = os.path.basename(parser.filename)
 
         # Stop conditions
@@ -378,7 +381,8 @@ def execute_pyccel(fname, *,
     except NotImplementedError as error:
         msg = str(error)
         errors.report(msg+'\n'+PYCCEL_RESTRICTION_TODO,
-            severity='error')
+            severity='error',
+            traceback=error.__traceback__)
         handle_error('code generation (wrapping)')
         raise PyccelCodegenError(msg) from None
     except PyccelError: