+++ CHANGELOG.md
@@ -12,12 +12,18 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+-   #1453 : Fix error-level developer mode output.
 -   \[INTERNALS\] Fix string base class selection.
 
 ### Changed
 
 -   Updating `stdlib` files if they are modified not just accessed.
+-   `pyccel_clean` tool now deletes folders **starting with** `__pyccel__` and `__epyccel__`.
+-   Pyccel-generated folder names are dependent on `PYTEST_XDIST_WORKER` when running with `pytest-xdist`.
 -   \[INTERNALS\] Add class object to class function call arguments.
+-   \[INTERNALS\] In `ast.numpyext` rename `Shape` as `NumpyShape`, `NumpyArraySize` as `NumpySize`
+-   \[INTERNALS\] In `ast.internals` rename `NumpyArraySize` as `NumpySize`, create new `NumpyArraySize` w/out `index` argument
+-   \[INTERNALS\] Make `NumpySize` a factory class (which cannot be instantiated)
 
 ### Deprecated